- Removed DiscordIcon component and switched to the existing image asset to reduce duplication. - Moved googleplaybadge.svg to the correct assets location. - Removed the `links` prop from `CTASection` and updated usages to align with the related GitHub issue. - Added JSDoc to `CTASection`
