+++ frontend-next-migration/src/shared/assets/icons/google-play-badge.png
